Colocate Client types
